requirements.txt
streamlit
google-generativeai
PyMuPDF
fpdf
spacy

